Exception kezelés hozzáadása update, delete és get műveleteket ellátó metódusokhoz
- getById() metódus lecserélése findById()-ra, mivel az elsőt akkor érdemes használni ha biztosan tudjuk, hogy a keresett id szerepel az adatbázisban, míg a második esetében ezt nem tudjuk, így dobhatunk kivételt, amennyiben az ellenőrzés után nincs az id az adatbázisban.
- A megfelelő helyeken a findById() használata után if-ekkel ellenőrzöm, hogy a keresett id megtalálható-e az adatbázisban, ha nem kivételt dob a program.

+++ src/main/java/hu/wysio/training/vivi/wysiokocsma/service/ItalService.java
@@ -2,12 +2,16 @@ package hu.wysio.training.vivi.wysiokocsma.service;
 
 import hu.wysio.training.vivi.wysiokocsma.converter.ItalConverter;
 import hu.wysio.training.vivi.wysiokocsma.dto.ItalDto;
+import hu.wysio.training.vivi.wysiokocsma.exception.ItalException;
+import hu.wysio.training.vivi.wysiokocsma.exception.WysioKocsmaException;
+import hu.wysio.training.vivi.wysiokocsma.model.ExceptionMessage;
 import hu.wysio.training.vivi.wysiokocsma.model.Ital;
 import hu.wysio.training.vivi.wysiokocsma.repository.ItalRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 
 @Service
@@ -25,23 +29,30 @@ public class ItalService {
         return ital.getId();
     }
 
-    public Ital updateItal(Long id, ItalDto italDto) {
+    public Ital updateItal(Long id, ItalDto italDto) throws WysioKocsmaException {
+        Optional<Ital> ital = italRepository.findById(id);
+
+        if (ital.isEmpty()) {
+            throw new ItalException(ExceptionMessage.NINCS_ITAL);
+        }
         Ital updatedItal = italConverter.toEntity(italDto);
 
-        Ital ital = italRepository.getById(id);
+        ital.get().setNev(updatedItal.getNev());
+        ital.get().setAdagMennyisege(updatedItal.getAdagMennyisege());
+        ital.get().setAlkoholTartalom(updatedItal.getAlkoholTartalom());
 
-        ital.setNev(updatedItal.getNev());
-        ital.setAdagMennyisege(updatedItal.getAdagMennyisege());
-        ital.setAlkoholTartalom(updatedItal.getAlkoholTartalom());
-
-        return italRepository.save(ital);
+        return italRepository.save(ital.get());
     }
 
     public List<Ital> findAll() {
         return italRepository.findAll();
     }
 
-    public void deleteItal(Long italId) {
+    public void deleteItal(Long italId) throws WysioKocsmaException {
+        if (italRepository.findById(italId).isEmpty()) {
+            throw new ItalException(ExceptionMessage.NINCS_ITAL);
+        }
+
         italRepository.deleteById(italId);
     }
 }
